raspberry-pi/temp_monitor/requirements.txt
RPi.GPIO==0.7.1
RPLCD==1.3.1
urllib3==2.2.3
